Split monolithic static file serving logic
The goal is to avoid a large number of blocks being created as a result of each request. The initial "hasty" implementation created and called route blocks for every request, which was a bit overkill and wasteful.

The new implementation has only one route block being create on `init` and calls instance methods for various tasks. This results in a simplified, more readable and a more efficient logic.